orchestrator/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Snowflake Integration
snowflake-snowpark-python==1.11.1
snowflake-connector-python==3.6.0

# Configuration
python-dotenv==1.0.0

# HTTP Clients
requests==2.31.0
httpx==0.26.0

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0
